+++ app/Http/Controllers/UsersController.php
@@ -3,25 +3,29 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Models\User;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Session;
+use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Crypt;
+use App\Models\User;
+use App\Models\AdminMenuRole;
+use App\Http\Menus\PermissionMenu;
+use App\Models\Menus;
+use Spatie\Permission\Models\Role;
 class UsersController extends Controller
 {
-
     /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
+    * Create a new controller instance.
+    *
+    * @return void
+    */
+    public function __construct(Request $request)
     {
         $this->middleware('auth');
-        $this->middleware('admin');
-        // $this->middleware('user');
-
-
-
+        session(['role_menu_id' => $request->current_menu_id]);
+        session(['user_name_menu' => $request->current_menu_name]);
+        
     }
 
     /**
@@ -55,7 +59,9 @@ class UsersController extends Controller
      */
     public function create()
     {
-        return view('default.users.userCreate');
+        return view('default.users.userCreate',[
+            'roles'    => Role::all(),
+        ]);
     }
 
     /**
@@ -66,29 +72,34 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'name' => 'required',
-            'email' => 'required',
-            'password' => 'required',
+        $permission_action = new PermissionMenu();
+		$permission_menu = $permission_action->permission(6, null, "insert");
+		if($permission_menu['status'] == true){
+            $validatedData = $request->validate([
+                'name' => 'required',
+                'email' => 'required',
+                'password' => 'required',  
+            ]);
             
-        ]);
-        $user = new User();
-        $user->name = $request->input('name');
-        $user->email = $request->input('email');
-        // $user->password = Crypt::encryptString($request->input('password'));
-        $user->password = Hash::make($request->input('password'));
-        $user->image = $request->input('image');
-        $user->email_verified_at = $request->input('email_verified_at');
-        $user->menuroles = $request->input('role');
-        $user->remember_token = $request->input('remember_token');
-        // $user->assignRole('user');
-        $user->save();
-        $user->assignRole($request->input('role'));
+            $user = new User();
+            $user->name = $request->input('name');
+            $user->email = $request->input('email');
+            $user->password = Hash::make($request->input('password'));
+            $user->email_verified_at = $request->input('email_verified_at');
+            $user->menuroles = $request->input('role');
+            $user->remember_token = $request->input('remember_token');
+            $user->assignRole($request->input('role'));
+            $user->save();
+            
+            $request->session()->flash('message_success', $permission_menu['message']);
+            $you = auth()->user();
+            $users = User::all();
+            return view('default.users.usersList', compact('users', 'you'));
+        }else{
+            $request->session()->flash("message_fail", $permission_menu['message']);
+            return view('default.users.userCreate');
 
-        $request->session()->flash('message', 'Successfully created menu');
-        $you = auth()->user();
-        $users = User::all();
-        return view('default.users.usersList', compact('users', 'you'));
+        }
     }
 
     /**
@@ -97,14 +108,17 @@ class UsersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Request $request, $id)
     {
+        $role = auth()->user();
+        $current_role_id = $role['id'];
+        $role_permission = AdminMenuRole::where('amrRoleID', $current_role_id )->where( 'amrMenusID', session('role_menu_id') )->first();
         $user = User::find($id);
-        return view('default.users.userEditForm', compact('user'));
+        return view('default.users.userEditForm', compact('user', 'role_permission'));
     }
     public function cropProfile(Request $request)
     {
-        $folderPath = public_path('uploads/');
+        $folderPath = public_path('uploads/users/');
         $image_parts = explode(";base64,", $request->profileImage);
         // return response()->json(['data'=>$image_parts]);
 
@@ -126,22 +140,30 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // $validatedData = $request->validate([
-        //     'name'       => 'required|min:1|max:256',
-        //     'email'      => 'required|email|max:256'
-        // ]);
-        $user = User::find($id);
-        $user->name       = $request->name;
-        $user->email      = $request->email;
-        $user->password      = Hash::make($request->password);
-        if($request->image !=""){
-            $user->image      = $request->image;
-        }
-        $user->save();
-        return response()->json(['success'=>$user]);
+        $permission_action = new PermissionMenu();
+        $permission_menu = $permission_action->permission(6, null,"update");
+        if($permission_menu['status'] == true){
+            $user = User::find($id);
+            $user->name          = $request->name;
+            $user->email         = $request->email;
+            if($user->password != $request->password){
+                $user->password      = $request->password;
 
-        // $request->session()->flash('message', 'Successfully updated user');
-        // return redirect()->route('users.index');
+            }else{
+                $user->password      = $user->password;
+
+            }
+            if($request->image !=""){
+                $user->image      = $request->image;
+            }
+            $user->save();
+            $request->session()->flash("message_success", $permission_menu['message']);
+			return response()->json($permission_menu['status']);
+
+		}else{
+			$request->session()->flash("message_fail", $permission_menu['message']);
+			return response()->json($permission_menu['status']);
+		}  
     }
     /**
      * Remove the specified resource from storage.
@@ -149,12 +171,20 @@ class UsersController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $user = User::find($id);
-        if($user){
-            $user->delete();
-        }
-        return redirect()->route('users.index');
+        $permission_action = new PermissionMenu();
+		$permission_menu = $permission_action->permission( 6, null,"delete");
+		if($permission_menu['status'] == true){
+            $user = User::find($id);
+            if($user){
+                $user->delete();
+            }
+            $request->session()->flash("message_success", $permission_menu['message']);
+			return redirect()->route('users.index'); 
+        }else{
+			$request->session()->flash("message_fail", $permission_menu['message']);
+			return redirect()->route('users.index'); 
+		}
     }
 }
